stats: Record ports separately and merge messages for stats
Separate port records allows independent recording in the future. It's
more flexible so we are not limited to sending a Sync on one port and
waiting for the RX on the other like its implemented now.

The idea is to more closely mimic the PTP daemon, since certain
messages like pdelay/delays requires being reactive and we don't want
to be sleeping for hundreds of milliseconds.

Signed-off-by: Casper Andersson <casper.casan@gmail.com>