time-error: Add initial draft
The idea of this mode is to have a GM running (e.g. ptp4l) on the
same, or a synchronized PHC, as tstest. This is basically reinventing
the TC mode to a more flexible solution that may supersede it
eventually. This should be applicable to both Boundary and
Transparent.

It sends out Delay/Pdelay, responds to Pdelay, and just listens to and
records everything else.

Signed-off-by: Casper Andersson <casper.casan@gmail.com>